ultimi aggiornamenti

+++ app/Http/Controllers/Upr/ApartmentController.php
@@ -7,11 +7,14 @@ use Illuminate\Support\Facades\Auth;
 use App\Apartment;
 use App\Service;
 use App\Image;
+use App\View;
+use App\Http\Traits\GetViews;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 
 class ApartmentController extends Controller
 {
+	use GetViews;
 	/**
 	 * Display a listing of the resource.
 	 *
@@ -45,10 +48,10 @@ class ApartmentController extends Controller
 		$data = $request->all();
 
 		if (empty($data["cover_img"])) {
-			$path = null;
-		} else {
-			$path = Storage::disk('public')->put('images', $data['cover_img']);
-		}
+            $path = 'storage/images/asset/nophoto.png';
+        } else {
+            $path = 'storage/' . Storage::disk('public')->put('images', $data['cover_img']);
+        }
 
 		$idUser = Auth::user()->id;
 		$request->validate([
@@ -102,16 +105,28 @@ class ApartmentController extends Controller
 		if (empty($apartment)) {
 			abort('404');
 		}
-		// $apartment->views += 1;
-		// $apartment->update();
+		$idUser = Auth::user()->id;
+		if ($apartment->user->id != $idUser) {
+			return redirect()->route('apartment.show', $apartment);
+			}
 
-		return view('upr.apartments.show', compact('apartment'));
+
+			$data = [
+				'apartment' => $apartment,
+				'views' => $this->GetViews($apartment->id)
+			];
+			return view('upr.apartments.show', $data);
 	}
 
 	public function show2(Apartment $apartment)
 	{
 		if (empty($apartment)) {
 			abort('404');
+		}
+
+		$idUser = Auth::user()->id;
+		if ($apartment->user->id != $idUser) {
+			return redirect()->route('apartment.show', $apartment);
 		}
 
 		$images = Image::all();
@@ -132,11 +147,13 @@ class ApartmentController extends Controller
 	 */
 	public function edit(Apartment $apartment)
 	{
-
+		$idUser = Auth::user()->id;
 		if (empty($apartment)) {
-			abort('404');
+			abort('400');
 		}
-
+		if ($apartment->user->id != $idUser) {
+					abort(401);
+				}
 
 		$services = Service::all();
 		$data = [
@@ -164,11 +181,11 @@ class ApartmentController extends Controller
 
 		$idUser = Auth::user()->id;
 		if (empty($apartment)) {
-			abort(404);
+			abort(400);
 		}
 
 		if ($apartment->user->id != $idUser) {
-			abort(404);
+			abort(401);
 		}
 
 		$request->validate([
@@ -194,7 +211,7 @@ class ApartmentController extends Controller
 		$apartment->active = $data['active'];
 		$apartment->price = $data['price'];
 
-		$apartment->cover_img = $path;
+		$apartment->cover_img = 'storage/' . $path;
 
 		$updated = $apartment->update();
 
@@ -217,8 +234,31 @@ class ApartmentController extends Controller
 	 */
 	public function destroy(Apartment $apartment)
 	{
+		$idUser = Auth::user()->id;
+		if (empty($apartment)) {
+			abort(400);
+		}
+
+		if ($apartment->user->id != $idUser) {
+			abort(401);
+		}
 		$apartment->services()->detach();
 		$apartment->delete();
 		return redirect()->route('upr.apartments.index');
+	}
+
+
+	public function statistics(Apartment $apartment)
+	{
+		$idUser = Auth::user()->id;
+		if (empty($apartment)) {
+			return redirect()->route('home');
+		}
+
+		if ($apartment->user->id != $idUser) {
+			abort(401);
+		}
+
+		return view('upr.apartments.statistics', compact('apartment'));
 	}
 }
